requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
gunicorn==20.1.0
PyPDF2==3.0.1
openai==1.3.0
firebase-admin==5.3.0
google-cloud-storage==2.5.0
google-auth==2.6.2
Werkzeug==2.0.1
pdfplumber==0.7.6
httpx==0.25.0 